Created usePromptBeforeUnload hook
Created hook to simplify confirmation dialog that displays when you try to close/refresh a page with unsaved changes
