+++ packages/ui/src/utils/hooks/index.ts
@@ -1,6 +1,7 @@
 export * from './useHistoryState';
-export * from './usePress';
 export * from './usePinchZoom';
+export * from './usePress';
+export * from './usePromptBeforeUnload';
 export * from './useReactHash';
 export * from './useReactPath';
 export * from './useReactSearch';
